requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.8.5
matplotlib>=3.7.0
numpy>=1.24.0